backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
flask-jwt-extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
razorpay==1.4.1
gunicorn==21.2.0
marshmallow==3.20.2
Pillow==10.2.0
cloudinary==1.39.0
